requirements.txt
numpy
pandas
sklearn
torch
torchvision
requests
atoma
pdfminer.six
